+++ src/main/java/com/teamtreehouse/techdegrees/dao/TodoDao.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface TodoDao {
     void add(Todo todo) throws DaoException;
     List<Todo> findAll();
-    Todo findById(int id);
+    Todo findById(long id);
 }
